Added Political Events and Altered African Scramble
Two events are added for after a communist or fascist dictatorship take
control and are able to reduce militancy and get their people their
needs. For this they get a reduction in militancy for most ideologies
and an increase in communists/fascists. Also I removed the line in the
African Scramble which made it very difficult for the Japanese or
Americans from taking part.