Massive Flag Revisions - Part 1 of ?
Included improved flags for many, many nations.  May upload a part 2 at
some point in the future, to replace many of the generic flags of
Indian/German minors.

Also included a couple of changes in country history files, to associate
certain flags with specific government types. (Japan will use the IJN
flag as an absolute monarchy as well as a fascist dictatorship, France
will use Ancien Regime flag only as an absolute monarchy. The latter may
be changed if a Bonapartiste government type is ever implemented.)